workflows: drop 24.3 Emacs from the testing matrix
Its byte-compilation seems to be broken, because it prints warnings
about unused values that seem to be completely spurious, the code it
complains about does not even have a let binding to begin with.

With that said, the testing matrix already has 24.5, so there's little
point in additionally testing an older Emacs 24.